apps/backend/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.2.1
gunicorn>=21.2.0,<22.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.13.0
cryptography
python-multipart
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
black
requests
